// src/modules/release/v1/interfaces/release_details.interface.ts
import { LinksStyle } from 'src/common/transformers/link_style.interface';

export interface ReleaseDetails {
  id: string;
  key: string;
  summary: string;
  lTitle: string;
  lSummary: string;
  lDescription: string;
  lSourceLinks: LinksStyle;
  lBusinessImpact: string;
  lSystemImpact: string;
  configSteps: string;
  userGuide: string;
  productLine: string;
  applicationArea: string;
  country: string;
  spt: string;
  locality: string;
  roadmapGroup: string;
}